cloud_function/requirements.txt
functions-framework==3.*
google-cloud-vision==3.5.0
google-cloud-storage==2.14.0
supabase==2.3.0
python-dotenv==1.0.0
